+++ R/label.R
@@ -80,10 +80,12 @@ xportr_label <- function(.df,
 
   ## End of common section
 
+
   assert_data_frame(.df)
   assert_string(domain, null.ok = TRUE)
   assert_metadata(metadata)
   assert_choice(verbose, choices = .internal_verbose_choices)
+  .df <- group_data_check(.df, verbose = verbose)
 
   domain_name <- getOption("xportr.domain_name")
   variable_name <- getOption("xportr.variable_name")