+++ R/length.R
@@ -83,13 +83,13 @@ xportr_length <- function(.df,
   verbose <- verbose %||%
     attr(.df, "_xportr.df_verbose_") %||%
     getOption("xportr.length_verbose", "none")
-
   ## End of common section
 
   assert_data_frame(.df)
   assert_string(domain, null.ok = TRUE)
   assert_metadata(metadata)
   assert_choice(verbose, choices = .internal_verbose_choices)
+  .df <- group_data_check(.df, verbose = verbose)
 
   domain_name <- getOption("xportr.domain_name")
   variable_length <- getOption("xportr.length")