+++ R/order.R
@@ -81,11 +81,11 @@ xportr_order <- function(.df,
     getOption("xportr.order_verbose", "none")
 
   ## End of common section
-
   assert_data_frame(.df)
   assert_string(domain, null.ok = TRUE)
   assert_metadata(metadata)
   assert_choice(verbose, choices = .internal_verbose_choices)
+  .df <- group_data_check(.df, verbose = verbose)
 
   domain_name <- getOption("xportr.domain_name")
   order_name <- getOption("xportr.order_name")