+++ R/write.R
@@ -63,8 +63,9 @@ xportr_write <- function(.df,
   # by the user.
 
   ## End of common section
-
+  .df <- group_data_check(.df)
   assert_data_frame(.df)
+
   assert_string(path)
   assert_numeric(max_size_gb, null.ok = TRUE)
   assert_metadata(metadata, null.ok = TRUE)
@@ -169,7 +170,7 @@ get_split_path <- function(path, ind) {
 export_to_xpt <- function(.df, path, max_size_gb, file_prefix) {
   # Convert GB to bytes
   max_size_bytes <- max_size_gb * 1000^3
-
+  .df <- group_data_check(.df)
   temp_file <- tempfile()
   write_xpt(.df, temp_file)